refactor: restructure loops and update function naming for consistency
- Relocated loop modules to appropriate `features` subdirectories (e.g., `location.ts` to `features/location-logger.ts`).
- Renamed functions for consistency with naming conventions (e.g., `load_location_logger` to `loadLocationLogger`).
- Consolidated headgear list into a `HEAD_GEAR` constant for cleaner code.
- Removed unnecessary debug console logs from loops.
